update phase_portraits with _matplotlib_points
