+++ dynamo/plot/scatters.py
@@ -1,29 +1,17 @@
 # code adapted from https://github.com/lmcinnes/umap/blob/7e051d8f3c4adca90ca81eb45f6a9d1372c076cf/umap/plot.py
 from ..configuration import _themes
 from .utils import despline, set_spine_linewidth, scatter_with_colorbar, scatter_with_legend, _select_font_color
+from .utils import is_gene_name, is_cell_anno_column, is_list_of_lists
+from .utils import _matplotlib_points, _datashade_points
+
 from ..tools.utils import get_mapper
 
 import numpy as np
 import pandas as pd
 from scipy.sparse import issparse
 
-from warnings import warn
-
 import matplotlib.colors
 import matplotlib.cm
-from matplotlib.patches import Patch
-import matplotlib.pyplot as plt
-
-import datashader as ds
-import datashader.transfer_functions as tf
-import datashader.bundling as bd
-
-import bokeh.plotting as bpl
-import bokeh.transform as btr
-from bokeh.plotting import output_notebook, output_file, show
-
-import holoviews as hv
-import holoviews.operation.datashader as hd
 
 
 def scatters(adata, genes, x=0, y=1, theme=None, type='expression', velocity_key='S', ekey='X', basis='umap', n_columns=1, \
@@ -86,10 +74,10 @@ def scatters(adata, genes, x=0, y=1, theme=None, type='expression', velocity_key
     -------
         Nothing but a scatter plot of cells.
     """
-    mapper = get_mapper()
 
     import matplotlib.pyplot as plt
     import seaborn as sns
+    mapper = get_mapper()
 
     point_size = 500.0 / np.sqrt(adata.shape[0]) if pointsize is None else 500.0 / np.sqrt(adata.shape[0]) * pointsize
     scatter_kwargs = dict(alpha=0.4, s=point_size, edgecolor=None, linewidth=0) # (0, 0, 0, 1)
@@ -435,4 +423,272 @@ def scatters(adata, genes, x=0, y=1, theme=None, type='expression', velocity_key
     plt.tight_layout()
     plt.show()
 
+
+def points(
+        adata,
+        x=0,
+        y=1,
+        color=None,
+        basis='umap',
+        layer='X',
+        highlights=None,
+        labels=None,
+        values=None,
+        theme=None,
+        cmap=None,
+        color_key=None,
+        color_key_cmap=None,
+        background="black",
+        ncols=1,
+        pointsize=None,
+        figsize=(7,5),
+        show_legend=True,
+        use_smoothed=True,
+        ax=None,
+        **kwargs):
+    """Plot an embedding as points. Currently this only works
+    for 2D embeddings. While there are many optional parameters
+    to further control and tailor the plotting, you need only
+    pass in the trained/fit umap model to get results. This plot
+    utility will attempt to do the hard work of avoiding
+    overplotting issues, and make it easy to automatically
+    colour points by a categorical labelling or numeric values.
+    This method is intended to be used within a Jupyter
+    notebook with ``%matplotlib inline``.
+    Parameters
+    ----------
+    adata: an anndata object.
+    basis: `str`
+        The reduced dimension.
+    labels: array, shape (n_samples,) (optional, default None)
+        An array of labels (assumed integer or categorical),
+        one for each data sample.
+        This will be used for coloring the points in
+        the plot according to their label. Note that
+        this option is mutually exclusive to the ``values``
+        option.
+    values: array, shape (n_samples,) (optional, default None)
+        An array of values (assumed float or continuous),
+        one for each sample.
+        This will be used for coloring the points in
+        the plot according to a colorscale associated
+        to the total range of values. Note that this
+        option is mutually exclusive to the ``labels``
+        option.
+    theme: string (optional, default None)
+        A color theme to use for plotting. A small set of
+        predefined themes are provided which have relatively
+        good aesthetics. Available themes are:
+           * 'blue'
+           * 'red'
+           * 'green'
+           * 'inferno'
+           * 'fire'
+           * 'viridis'
+           * 'darkblue'
+           * 'darkred'
+           * 'darkgreen'
+    cmap: string (optional, default 'Blues')
+        The name of a matplotlib colormap to use for coloring
+        or shading points. If no labels or values are passed
+        this will be used for shading points according to
+        density (largely only of relevance for very large
+        datasets). If values are passed this will be used for
+        shading according the value. Note that if theme
+        is passed then this value will be overridden by the
+        corresponding option of the theme.
+    color_key: dict or array, shape (n_categories) (optional, default None)
+        A way to assign colors to categoricals. This can either be
+        an explicit dict mapping labels to colors (as strings of form
+        '#RRGGBB'), or an array like object providing one color for
+        each distinct category being provided in ``labels``. Either
+        way this mapping will be used to color points according to
+        the label. Note that if theme
+        is passed then this value will be overridden by the
+        corresponding option of the theme.
+    color_key_cmap: string (optional, default 'Spectral')
+        The name of a matplotlib colormap to use for categorical coloring.
+        If an explicit ``color_key`` is not given a color mapping for
+        categories can be generated from the label list and selecting
+        a matching list of colors from the given colormap. Note
+        that if theme
+        is passed then this value will be overridden by the
+        corresponding option of the theme.
+    background: string (optional, default 'white)
+        The color of the background. Usually this will be either
+        'white' or 'black', but any color name will work. Ideally
+        one wants to match this appropriately to the colors being
+        used for points etc. This is one of the things that themes
+        handle for you. Note that if theme
+        is passed then this value will be overridden by the
+        corresponding option of the theme.
+    width: int (optional, default 800)
+        The desired width of the plot in pixels.
+    height: int (optional, default 800)
+        The desired height of the plot in pixels
+    show_legend: bool (optional, default True)
+        Whether to display a legend of the labels
+    Returns
+    -------
+    result: matplotlib axis
+        The result is a matplotlib axis with the relevant plot displayed.
+        If you are using a notbooks and have ``%matplotlib inline`` set
+        then this will simply display inline.
+    """
+
+    import matplotlib.pyplot as plt
+    import seaborn as sns
+
+    x, y = x[0], y[0]
+
+    if use_smoothed: mapper = get_mapper()
+
+
+    n_c, n_l, n_b = 0 if color is None else len(color), 0 if layer is None else len(layer), 0 if basis is None else len(basis)
+
+    point_size = 500.0 / np.sqrt(adata.shape[0]) if pointsize is None else 500.0 / np.sqrt(adata.shape[0]) * pointsize
+    scatter_kwargs = dict(alpha=0.4, s=point_size, edgecolor=None, linewidth=0) # (0, 0, 0, 1)
+    if kwargs is not None:
+        scatter_kwargs.update(kwargs)
+
+    if all([i in adata.layers.keys() for i in ['X_new', 'X_total']]):
+        mode = 'labeling'
+    elif all([i in adata.layers.keys() for i in ['X_spliced', 'X_unspliced']]):
+        mode = 'splicing'
+    elif all([i in adata.layers.keys() for i in ['X_uu', 'X_ul', 'X_su', 'X_sl']]):
+        mode = 'full'
+    else:
+        raise Exception('your data should be in one of the proper mode: labelling (has X_new/X_total layers), splicing '
+                        '(has X_spliced/X_unspliced layers) or full (has X_uu/X_ul/X_su/X_sl layers)')
+
+    font_color = _select_font_color(background)
+    if background == 'black':
+        # https://github.com/matplotlib/matplotlib/blob/master/lib/matplotlib/mpl-data/stylelib/dark_background.mplstyle
+        sns.set(rc={'axes.facecolor': background, 'axes.edgecolor': background, 'figure.facecolor': background, 'figure.edgecolor': background,
+                    'axes.grid': False, "ytick.color": "w", "xtick.color": "w", "axes.labelcolor": "w", "axes.edgecolor": "w",
+                    "savefig.facecolor": 'k', "savefig.edgecolor": 'k', "grid.color": 'w', "text.color": font_color,
+                    "lines.color": 'w', "patch.edgecolor": 'w', 'figure.edgecolor': 'w',
+                    })
+    else:
+        sns.set(rc={'axes.facecolor': background, 'figure.facecolor': background, "text.color": font_color, 'axes.grid': False})
+
+    total_panels, ncols = n_c * n_l * n_b, min(n_c, ncols)
+    nrow, ncol = int(np.ceil(total_panels / ncols)), ncols
+    if figsize is None: figsize = plt.rcParams['figsize']
+
+    font_color = _select_font_color(background)
+    if background == 'black':
+        # https://github.com/matplotlib/matplotlib/blob/master/lib/matplotlib/mpl-data/stylelib/dark_background.mplstyle
+        sns.set(rc={'axes.facecolor': background, 'axes.edgecolor': background, 'figure.facecolor': background, 'figure.edgecolor': background,
+                    'axes.grid': False, "ytick.color": font_color, "xtick.color": font_color, "axes.labelcolor": font_color, "axes.edgecolor": font_color,
+                    "savefig.facecolor": 'k', "savefig.edgecolor": 'k', "grid.color": font_color, "text.color": font_color,
+                    "lines.color": font_color, "patch.edgecolor": font_color, 'figure.edgecolor': font_color,
+                    })
+    else:
+        sns.set(rc={'axes.facecolor': background, 'figure.facecolor': background, 'axes.grid': False})
+
+    if total_panels > 1:
+        plt.figure(None, (figsize[0] * ncol, figsize[1] * nrow), facecolor=background)
+        gs = plt.GridSpec(nrow, ncol)
+
+    i = 0
+    for cur_b in basis:
+        for cur_l in layer:
+            if use_smoothed: cur_l_smoothed = mapper[cur_l]
+            prefix = cur_l + '_'
+
+            if prefix + cur_b in adata.obsm.keys():
+                x_, y_ = adata.obsm[prefix + cur_b][:, int(x)], adata.obsm[prefix + cur_b][:, int(y)]
+            else:
+                continue
+            for cur_c in color:
+                if cur_l in ['protein', 'X_protein']:
+                    _color = adata.obsm[cur_l].loc[cur_c, :]
+                else:
+                    _color = adata.obs_vector(cur_c, layer=cur_l)
+
+                if type(x) is int and type(y) is int:
+                    points = pd.DataFrame({cur_b + '_0': adata.obsm[prefix + cur_b][:, x], \
+                                              cur_b + '_1': adata.obsm[prefix + cur_b][:, y]})
+                    points.columns = [cur_b + '_1', cur_b + '_2']
+                elif is_gene_name(x) and is_gene_name(y):
+                    points = pd.DataFrame({x: adata.obs_vector(x, cur_l_smoothed), \
+                                              y: adata.obs_vector(y, cur_l_smoothed)})
+                    points.columns = [x + ' (' + cur_l_smoothed + ')', y + ' (' + cur_l_smoothed + ')']
+                elif is_cell_anno_column(x) and is_gene_name(y):
+                    points = pd.DataFrame({x: adata.obs_vector(x), \
+                                              y: adata.obs_vector(y, cur_l_smoothed)})
+                    points.columns = [x, y + ' (' + cur_l_smoothed + ')']
+
+                is_not_continous = _color.dtype.name == 'category'
+
+                if is_not_continous:
+                    labels = _color
+                    if theme is None: theme = 'glasbey_dark'
+                else:
+                    values = _color
+                    if theme is None: theme = 'inferno' if cur_l is not 'velocity' else 'div_blue_red'
+
+                if cmap is None: cmap = _themes[theme]["cmap"]
+                if color_key_cmap is None: color_key_cmap = _themes[theme]["color_key_cmap"]
+                if background is None: background = _themes[theme]["background"]
+
+                if labels is not None and values is not None:
+                    raise ValueError(
+                        "Conflicting options; only one of labels or values should be set"
+                    )
+
+                points = adata.obsm['X_' + basis]
+
+                if total_panels > 1:
+                    ax = plt.subplot(gs[i])
+                i += 1
+
+                # if highligts is a list of lists - each list is relate to each color element
+                if is_list_of_lists(highlights):
+                    _highlights = highlights[color.index(cur_c)]
+                    _highlights = _highlights if all([i in _color for i in _highlights]) else None
+                else:
+                    _highlights = highlights if all([i in _color for i in highlights]) else None
+
+                if points.shape[0] <= figsize[0] * figsize[1] * 1000:
+                    ax = _matplotlib_points(
+                        points,
+                        ax,
+                        labels,
+                        values,
+                        highlights,
+                        cmap,
+                        color_key,
+                        color_key_cmap,
+                        background,
+                        figsize[0],
+                        figsize[1],
+                        show_legend
+                    )
+                else:
+                    ax = _datashade_points(
+                        points,
+                        ax,
+                        labels,
+                        values,
+                        highlights,
+                        cmap,
+                        color_key,
+                        color_key_cmap,
+                        background,
+                        figsize[0],
+                        figsize[1],
+                        show_legend
+                    )
+
+                ax.set_xlabel(points.columns[0])
+                ax.set_ylabel(points.columns[1])
+                ax.set_title(cur_c)
+
+    plt.tight_layout()
+    plt.show()
+
+    # dyn.configuration.reset_rcParams()
+    if total_panels == 1: return ax
